# Conflicts:
#	lib/my/src/headers/my.h
#	lib/my/src/my_strlen.c
#	lib/my/src/my_strstr.c
#	src/headers/string.h
#	src/main.c

+++ lib/my/src/my_strlen.c
@@ -1,6 +1,6 @@
 #include <stddef.h>
 
-int       my_strlen(const char *str)
+size_t    my_strlen(const char *str)
 {
   size_t  len;
